All DAOs implemented and configured.

+++ src/main/java/com/google/code/imazon/model/book/BookDao.java
@@ -2,7 +2,7 @@ package com.google.code.imazon.model.book;
 
 import java.util.List;
 
-import es.udc.pojo.modelutil.dao.*;
+import es.udc.pojo.modelutil.dao.GenericDao;
 import es.udc.pojo.modelutil.exceptions.InstanceNotFoundException;
 
 public interface BookDao extends GenericDao<Book, Long> {
@@ -10,16 +10,16 @@ public interface BookDao extends GenericDao<Book, Long> {
 	public List<Book> findBooksByCategoryAndKeys(Short categoryId, String keys,
 			int start, int count) throws InstanceNotFoundException;
 	
-	public int getNumberOfBooksByCategoryAndKeys(Short categoryId, String keys);
+	public Long getNumberOfBooksByCategoryAndKeys(Short categoryId, String keys);
 
 	public List<Book> findBooks(int start, int count)
 			throws InstanceNotFoundException;
 	
-	public int getNumberOfBooks();
+	public Long getNumberOfBooks();
 
 	public List<Book> findBooksByPublisherId(Long publisherId, int start, int count)
 			throws InstanceNotFoundException;
 
-	public int getNumberOfBooksByPublisherId(Long publisherId);
+	public Long getNumberOfBooksByPublisherId(Long publisherId);
 
 }
